+++ src/main/java/com/horus/projeto/dto/ProducaoRequestDTO.java
@@ -3,6 +3,7 @@ package com.horus.projeto.dto;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -11,5 +12,5 @@ public class ProducaoRequestDTO {
     /** ID do produto a ser produzido (deve ser PF ou MPPF). */
     private Long codProduto;
     /** Quantidade de unidades a produzir. */
-    private Integer quantidade;
+    private BigDecimal quantidade;
 }
